requirements.txt
pyTelegramBotAPI==4.15.4
Pillow==10.2.0 